create basket

+++ lib/repositories/item_repository.rb
@@ -32,7 +32,7 @@ class ItemRepository
   def load_csv
     CSV.foreach(@csv_file, headers: :first_row, header_converters: :symbol) do |row|
       row[:productcode] = row[:productcode].to_i
-      row[:price] = row[:price].to_i
+      row[:price] = row[:price].to_f
       @items << Item.new(row)
     end
       @next_productcode = @items.empty? ? 1 : @items.last.productcode + 1
